requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asteroid-filterbanks==0.4.0
attrs==25.4.0
audioread==3.1.0
boto3==1.42.62
botocore==1.42.62
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
coverage==7.13.4
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
einops==0.8.1
execnet==2.1.2
fastapi==0.128.0
ffmpeg-python==0.2.0
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
librt==0.8.1
lightning==2.6.0
lightning-utilities==0.15.2
llama_cpp_python_binary==0.2.26
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.8.2
primePy==1.3
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyannote-audio==4.0.3
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
pytest-cov==7.0.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.45
sse-starlette==3.1.2
starlette==0.50.0
starlette-context==0.3.6
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchcodec==0.7.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.1
uvicorn==0.40.0
yarl==1.22.0
yt-dlp==2026.2.4
zipp==3.23.0
